pyproject.toml
[project]
name = "meeting-transcription-agent"
version = "0.1.0"
description = "M365 Copilot custom engine agent: fetch meeting transcriptions via Graph and summarize with Azure OpenAI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "microsoft-agents-hosting-aiohttp>=0.7.0",
    "microsoft-agents-hosting-teams>=0.7.0",
    "microsoft-agents-authentication-msal>=0.7.0",
    "msal>=1.24.0",
    "httpx>=0.25.0",
    "openai>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meeting_agent"]

[tool.uv]
dev-dependencies = []
